Create additional bundle class according to Symfony best practices
Symfony Flex registers bundle automatically but instead of using
`FormHandlerBundle` is goes with `HostnetFormHandlerBundle`
