+++ src/HostnetFormHandlerBundle.php
@@ -1,0 +1,28 @@
+<?php
+namespace Hostnet\Bundle\FormHandlerBundle;
+
+use Hostnet\Bundle\FormHandlerBundle\DependencyInjection\Compiler\FormParamConverterCompilerPass;
+use Symfony\Component\Config\FileLocator;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Loader\YamlFileLoader;
+use Symfony\Component\HttpKernel\Bundle\Bundle;
+
+/**
+ * @author Yannick de Lange <ydelange@hostnet.nl>
+ * @author Iltar van der Berg <ivanderberg@hostnet.nl>
+ */
+class HostnetFormHandlerBundle extends Bundle
+{
+    /**
+     * @see \Symfony\Component\HttpKernel\Bundle\Bundle::build()
+     */
+    public function build(ContainerBuilder $container)
+    {
+        // load default services.yml
+        $loader = new YamlFileLoader($container, new FileLocator(__DIR__ . '/Resources/config'));
+        $loader->load('services.yml');
+
+        parent::build($container);
+        $container->addCompilerPass(new FormParamConverterCompilerPass());
+    }
+}
